+++ NEWS.md
@@ -16,6 +16,8 @@ Language changes
   * Big integer literals and command syntax (backticks) are now parsed with the name of
     the macro (`@int128_str`, `@uint128_str`, `@big_str`, `@cmd`) qualified to refer
     to the `Core` module ([#29968]).
+  * Using the same name for both a local variable and a static parameter is now an error instead
+    of a warning ([#29429]).
 
 New library functions
 ---------------------